+++ core/persistence/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/persistence/data/FavoriteCommunityModel.kt
@@ -2,5 +2,5 @@ package com.github.diegoberaldin.raccoonforlemmy.core.persistence.data
 
 data class FavoriteCommunityModel(
     val id: Long? = null,
-    val communityId: Int? = null,
+    val communityId: Long? = null,
 )
