+++ domain/lemmy/data/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/domain/lemmy/data/PrivateMessageModel.kt
@@ -1,7 +1,7 @@
 package com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data
 
 data class PrivateMessageModel(
-    val id: Int = 0,
+    val id: Long = 0,
     val content: String? = null,
     val creator: UserModel? = null,
     val recipient: UserModel? = null,
@@ -10,7 +10,7 @@ data class PrivateMessageModel(
     val read: Boolean = false,
 )
 
-fun PrivateMessageModel.otherUser(currentUserId: Int): UserModel? = when (currentUserId) {
+fun PrivateMessageModel.otherUser(currentUserId: Long): UserModel? = when (currentUserId) {
     creator?.id -> recipient
     recipient?.id -> creator
     else -> null
